+++ layouts/proartp16.py
@@ -8,44 +8,46 @@ center_button_diameter = 364
 circle_center_x = 586
 circle_center_y = 573
 
+# current_value and title is optional because it is used for the user interface
 app_shortcuts = {
-    "/usr/share/code/code": {
+    "not_implemented_yet": { # "/usr/share/code/code"
+        "Volume": {
+            "icon": "foo",
+            "clockwise": [
+              # works even better with `dconf write /org/gnome/desktop/sound/allow-volume-above-100-percent true`
+              {"key": EV_KEY.KEY_VOLUMEUP, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT, "current_value": "pactl get-sink-volume @DEFAULT_SINK@ | grep -oP '\d+%' | head -n 1 | tr -d '%'", "slices_count": 10, "title": "Volume"}
+            ],
+            "counterclockwise": [
+              # works even better with `dconf write /org/gnome/desktop/sound/allow-volume-above-100-percent true`
+              {"key": EV_KEY.KEY_VOLUMEDOWN, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT, "current_value": "pactl get-sink-volume @DEFAULT_SINK@ | grep -oP '\d+%' | head -n 1 | tr -d '%'", "slices_count": 10, "title": "Volume"}
+            ]
+        },
+        "Scroll": {
+            "icon": "bar",
+            "clockwise": [
+               {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [1, 120], "trigger": "immediate"},
+            ],
+            "counterclockwise": [
+              {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [-1, -120], "trigger": "immediate"},
+            ]
+        }
+    },
+    "firefox": {
+        # ... e.g. same as below (single function mode) or above (multi-function mode: not implemented yet)...
+    },
+    "none": {
         "center": [
           {"key": EV_KEY.KEY_MUTE, "trigger": "release", "duration": 1, "modifier": EV_KEY.KEY_LEFTSHIFT}
         ],
         "clockwise": [
-          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [1, 120], "trigger": "immediate"},
-          {"key": EV_KEY.KEY_VOLUMEUP, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT}
+          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [1, 120], "trigger": "immediate", "title": "Scroll"},
+          # works even better with `dconf write /org/gnome/desktop/sound/allow-volume-above-100-percent true`
+          {"key": EV_KEY.KEY_VOLUMEUP, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT, "current_value": "pactl get-sink-volume @DEFAULT_SINK@ | grep -oP '\d+%' | head -n 1 | tr -d '%'", "slices_count": 10, "title": "Volume"}
         ],
         "counterclockwise": [
-          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [-1, -120], "trigger": "immediate"},
-          {"key": EV_KEY.KEY_VOLUMEDOWN, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT}
-        ]
-    },
-    "firefox": {
-        "center": [
-          {"key": EV_KEY.KEY_MUTE, "trigger": "release", "duration": 1},
-        ],
-        "clockwise": [
-          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [1, 120], "trigger": "immediate"},
-          {"key": EV_KEY.KEY_VOLUMEUP, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT}
-        ],
-        "counterclockwise": [
-          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [-1, -120], "trigger": "immediate"},
-          {"key": EV_KEY.KEY_VOLUMEDOWN, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT}
-        ]
-    },
-    "none": {
-        "center": [
-          {"key": EV_KEY.KEY_MUTE, "trigger": "release", "duration": 1},
-        ],
-        "clockwise": [
-          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [1, 120], "trigger": "immediate"},
-          {"key": EV_KEY.KEY_VOLUMEUP, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT}
-        ],
-        "counterclockwise": [
-          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [-1, -120], "trigger": "immediate"},
-          {"key": EV_KEY.KEY_VOLUMEDOWN, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT}
+          {"key": [EV_REL.REL_WHEEL, EV_REL.REL_WHEEL_HI_RES], "value": [-1, -120], "trigger": "immediate", "title": "Scroll"},
+          # works even better with `dconf write /org/gnome/desktop/sound/allow-volume-above-100-percent true`
+          {"key": EV_KEY.KEY_VOLUMEDOWN, "trigger": "immediate", "modifier": EV_KEY.KEY_LEFTSHIFT, "current_value": "pactl get-sink-volume @DEFAULT_SINK@ | grep -oP '\d+%' | head -n 1 | tr -d '%'", "slices_count": 10, "title": "Volume"}
         ]
     }
 }